+++ data-raw/sysdata.R
@@ -1,8 +1,9 @@
-# sysdata.rda generation file
+# sysdata.rda generation file (included params for tests)
 
 freq_go_pairs <- GAPGOM:::freq_go_pairs
 lncrnapred_baseresult <- GAPGOM:::lncrnapred_baseresult
 maintopo_baseresult <- GAPGOM:::maintopo_baseresult
 genetopo_baseresult <- GAPGOM:::genetopo_baseresult
+termtopo_baseresult <- GAPGOM:::termtopo_baseresult
 benchmarks <- GAPGOM:::benchmarks
-save(freq_go_pairs, lncrnapred_baseresult, maintopo_baseresult, genetopo_baseresult, benchmarks, file = "/media/casper/USB_ccpeters/Repositories/Personal/gapgom/R/sysdata.rda", compress = "xz", compression_level = 9)
+save(freq_go_pairs, lncrnapred_baseresult, maintopo_baseresult, genetopo_baseresult, termtopo_baseresult, benchmarks, file = "/media/casper/USB_ccpeters/Repositories/Personal/gapgom/R/sysdata.rda", compress = "xz", compression_level = 9)